update topic subscribe/unsubscribe

+++ lib/common/topic/topic.dart
@@ -20,9 +20,9 @@ class TopicCommon with Tag {
   StreamSink<TopicSchema> get _addSink => _addController.sink;
   Stream<TopicSchema> get addStream => _addController.stream;
 
-  StreamController<int> _deleteController = StreamController<int>.broadcast();
-  StreamSink<int> get _deleteSink => _deleteController.sink;
-  Stream<int> get deleteStream => _deleteController.stream;
+  StreamController<String> _deleteController = StreamController<String>.broadcast();
+  StreamSink<String> get _deleteSink => _deleteController.sink;
+  Stream<String> get deleteStream => _deleteController.stream;
 
   StreamController<TopicSchema> _updateController = StreamController<TopicSchema>.broadcast();
   StreamSink<TopicSchema> get _updateSink => _updateController.sink;
@@ -36,11 +36,18 @@ class TopicCommon with Tag {
     _updateController.close();
   }
 
+  checkAllTopics() async {
+    List<TopicSchema> topics = await queryList();
+    topics.forEach((TopicSchema topic) {
+      checkExpireAndSubscribe(topic.topic, subscribeFirst: false, emptyAdd: false);
+    });
+  }
+
   /// ***********************************************************************************************************
   /// *********************************************** subscribe *************************************************
   /// ***********************************************************************************************************
 
-  // caller = everyone
+  // caller = self
   Future<TopicSchema?> subscribe(String? topicName, {double fee = 0}) async {
     if (topicName == null || topicName.isEmpty) return null;
     // subscribe/expire
@@ -61,11 +68,11 @@ class TopicCommon with Tag {
       await subscriberCommon.refreshSubscribers(topicName, meta: false);
     }
 
-    await Future.delayed(Duration(seconds: 2));
+    await Future.delayed(Duration(seconds: 3));
     return _topic;
   }
 
-  // caller = everyone
+  // caller = self
   Future<TopicSchema?> checkExpireAndSubscribe(String? topicName, {bool subscribeFirst = false, bool emptyAdd = false, double fee = 0}) async {
     if (topicName == null || topicName.isEmpty) return null;
 
@@ -85,34 +92,50 @@ class TopicCommon with Tag {
     if (!exists.joined || exists.subscribeAt == null || exists.subscribeAt! <= 0 || exists.expireBlockHeight == null || exists.expireBlockHeight! <= 0) {
       if (exists.expireBlockHeight == null || exists.expireBlockHeight! <= 0) {
         if (!subscribeFirst) {
-          logger.i("$TAG - checkExpireAndSubscribe - can not subscribeFirst - topic:$exists");
+          logger.i("$TAG - checkExpireAndSubscribe - can not subscribeFirst - topic:$exists"); // TODO:GG 退订是不是0
           return null;
         }
       }
-      int subscribeAt = exists.subscribeAt ?? DateTime.now().millisecondsSinceEpoch;
       int expireHeight = await _getExpireAt(topicName, clientCommon.address);
       if (expireHeight > 0) {
-        // sync node info
-        bool success = await setJoined(exists.id, true, subscribeAt: subscribeAt, expireBlockHeight: expireHeight, notify: true);
-        if (success) {
-          exists.joined = true;
-          exists.subscribeAt = subscribeAt;
-          exists.expireBlockHeight = expireHeight;
+        int createAt = exists.createAt ?? DateTime.now().millisecondsSinceEpoch;
+        if ((DateTime.now().millisecondsSinceEpoch - createAt).abs() > 1000 * 60) {
+          logger.d("$TAG - checkExpireAndSubscribe - DB expire but node not expire - topic:$exists");
+          int subscribeAt = exists.subscribeAt ?? DateTime.now().millisecondsSinceEpoch;
+          bool success = await setJoined(exists.id, true, subscribeAt: subscribeAt, expireBlockHeight: expireHeight, notify: true);
+          if (success) {
+            exists.joined = true;
+            exists.subscribeAt = subscribeAt;
+            exists.expireBlockHeight = expireHeight;
+          }
+        } else {
+          logger.d("$TAG - checkExpireAndSubscribe - DB expire but node not expire, maybe in txPool - topic:$exists");
+          noSubscribed = true;
         }
       } else {
-        // no subscribe history
+        logger.d("$TAG - checkExpireAndSubscribe - no subscribe history - topic:$exists");
         noSubscribed = true;
+      }
+    } else {
+      int createAt = exists.createAt ?? DateTime.now().millisecondsSinceEpoch;
+      int expireHeight = await _getExpireAt(topicName, clientCommon.address);
+      if (expireHeight <= 0 && exists.joined && (DateTime.now().millisecondsSinceEpoch - createAt).abs() > 1000 * 60) {
+        logger.i("$TAG - checkExpireAndSubscribe - db no expire but node expire - topic:$exists");
+        bool success = await setJoined(exists.id, false, notify: true);
+        if (success) exists.joined = false;
+      } else {
+        logger.d("$TAG - checkExpireAndSubscribe - DB not expire but node expire, maybe in txPool - topic:$exists");
       }
     }
 
     // check expire
     int? globalHeight = await clientCommon.client?.getHeight();
-    if (noSubscribed || (await exists.shouldResubscribe(globalHeight: globalHeight))) {
+    if (noSubscribed) {
       bool subscribeSuccess;
       if (exists.isOwner(clientCommon.address)) {
         // private + owner
         SubscriberSchema? _subscriberMe = await subscriberCommon.onSubscribe(topicName, clientCommon.address, permPage: 0);
-        Map<String, dynamic> meta = await subscriberCommon.getPermissionsMetaByPage(topicName, _subscriberMe, appendPermPage: 0);
+        Map<String, dynamic> meta = await subscriberCommon.createMetaByDB(topicName, _subscriberMe, appendPermPage: 0);
         subscribeSuccess = await _clientSubscribe(
           topicName,
           height: Global.topicDefaultSubscribeHeight,
@@ -121,7 +144,7 @@ class TopicCommon with Tag {
           meta: meta,
         );
       } else {
-        // publish / private normal member
+        // publish / private normal member // TODO:GG 私有群普通成员不用meta吧
         subscribeSuccess = await _clientSubscribe(
           topicName,
           height: Global.topicDefaultSubscribeHeight,
@@ -178,7 +201,7 @@ class TopicCommon with Tag {
   /// ************************************************ members **************************************************
   /// ***********************************************************************************************************
 
-  // caller = owner
+  // caller = public / private + owner
   Future<SubscriberSchema?> invitee(String? topicName, String? clientAddress) async {
     if (topicName == null || topicName.isEmpty || clientAddress == null || clientAddress.isEmpty) return null;
     if (clientAddress == clientCommon.address) {
@@ -189,7 +212,7 @@ class TopicCommon with Tag {
     // topic owner
     TopicSchema? _topic = await queryByTopic(topicName);
     if (_topic == null) return null;
-    if (_topic.isOwner(clientCommon.address) != true) {
+    if (_topic.isPrivate && !_topic.isOwner(clientCommon.address)) {
       Toast.show(S.of(Global.appContext).member_no_auth_invite);
       return null;
     }
@@ -206,29 +229,32 @@ class TopicCommon with Tag {
     if (_msg == null) return null;
 
     // subscriber update
-    int appendPermPage = await subscriberCommon.queryMaxPermPageByTopic(topicName);
+    int? appendPermPage = _topic.isPrivate ? await subscriberCommon.queryMaxPermPageByTopic(topicName) : null;
     _subscriber = await subscriberCommon.onInvitedSend(topicName, clientAddress, permPage: appendPermPage);
     if (_subscriber == null) return null;
 
-    // client subscribe
-    Map<String, dynamic> meta = await subscriberCommon.getPermissionsMetaByPage(topicName, _subscriber, appendPermPage: appendPermPage);
-    bool subscribeSuccess = await _clientSubscribe(
-      topicName,
-      height: Global.topicDefaultSubscribeHeight,
-      fee: 0,
-      permissionPage: appendPermPage,
-      meta: meta,
-    );
-    if (subscribeSuccess) {
-      Toast.show(S.of(Global.appContext).invitation_sent);
-      return _subscriber;
-    } else {
-      await subscriberCommon.delete(_subscriber.id, notify: true);
-      return null;
+    if (_topic.isPrivate) {
+      // client subscribe
+      Map<String, dynamic> meta = await subscriberCommon.createMetaByDB(topicName, _subscriber, appendPermPage: appendPermPage);
+      bool subscribeSuccess = await _clientSubscribe(
+        topicName,
+        height: Global.topicDefaultSubscribeHeight,
+        fee: 0,
+        permissionPage: appendPermPage,
+        meta: meta,
+      );
+      if (subscribeSuccess) {
+        Toast.show(S.of(Global.appContext).invitation_sent);
+        return _subscriber;
+      } else {
+        await subscriberCommon.delete(_subscriber.id, notify: true);
+        return null;
+      }
     }
+    return _subscriber;
   }
 
-  // caller = owner
+  // caller = private + owner TODO:GG 缺个协议
   Future<SubscriberSchema?> kick(String? topicName, String? clientAddress) async {
     if (topicName == null || topicName.isEmpty || clientAddress == null || clientAddress.isEmpty) return null;
     if (clientAddress == clientCommon.address) return null;
@@ -236,7 +262,7 @@ class TopicCommon with Tag {
     // topic owner
     TopicSchema? _topic = await queryByTopic(topicName);
     if (_topic == null) return null;
-    if (_topic.isOwner(clientCommon.address) != true) return null;
+    if (!_topic.isOwner(clientCommon.address)) return null;
 
     // subscriber status
     SubscriberSchema? _subscriber = await subscriberCommon.queryByTopicChatId(topicName, clientAddress);
@@ -323,7 +349,7 @@ class TopicCommon with Tag {
   /// ********************************************** unsubscribe ************************************************
   /// ***********************************************************************************************************
 
-  // caller = everyone
+  // caller = self
   Future<TopicSchema?> unsubscribe(String? topicName, {double fee = 0, int? permissionPage, bool deleteDB = false}) async {
     if (topicName == null || topicName.isEmpty) return null;
 
@@ -343,7 +369,7 @@ class TopicCommon with Tag {
     if (deleteDB) await delete(exists?.id, notify: true);
     if (deleteDB) await subscriberCommon.deleteByTopic(topicName);
 
-    await Future.delayed(Duration(seconds: 2));
+    await Future.delayed(Duration(seconds: 3));
     return exists;
   }
 
@@ -381,16 +407,25 @@ class TopicCommon with Tag {
   // caller = everyone
   Future<bool> isJoined(String? topicName, String? clientAddress, {int? globalHeight}) async {
     if (topicName == null || topicName.isEmpty) return false;
+    TopicSchema? exists = await queryByTopic(topicName);
+    int createAt = exists?.createAt ?? DateTime.now().millisecondsSinceEpoch;
+    if (exists != null && (DateTime.now().millisecondsSinceEpoch - createAt).abs() < 1000 * 60) {
+      return exists.joined; // maybe in txPool
+    }
     int expireHeight = await _getExpireAt(topicName, clientAddress);
-    if (expireHeight <= 0) return false;
+    if (expireHeight <= 0) {
+      logger.d("$TAG - isJoined - expireHeight <= 0 - topicName:$topicName - clientAddress:$clientAddress");
+      return false;
+    }
     globalHeight = globalHeight ?? await clientCommon.client?.getHeight();
-    if (globalHeight == null || globalHeight <= 0) return false;
-    // TODO:GG perm??? 踢出去之后怎么保证她不能加进来
-    // TODO:GG (owner更新permission，node会自动处理subscribe？，然后所有的subscribers都走client，而不是DB)
+    if (globalHeight == null || globalHeight <= 0) {
+      logger.w("$TAG - isJoined - globalHeight <= 0 - topicName:$topicName");
+      return false;
+    }
     return expireHeight >= globalHeight;
   }
 
-  // caller = owner
+  // caller = private + owner TODO:GG 别人调用会有返回吗？
   Future refreshSubscribersByOwner(String? topicName, {bool allPermPage = false, int permPage = 0, int? maxPage}) async {
     if (topicName == null || topicName.isEmpty) return;
     Map<String, dynamic> meta = await _getMeta(topicName, permPage);
@@ -416,8 +451,7 @@ class TopicCommon with Tag {
     String indexWithPubKey = '__${permPage}__.__permission__.$ownerPubKey';
     Map<String, dynamic> result = await _clientGetSubscription(topicName, indexWithPubKey);
     if (result['meta']?.toString().isNotEmpty == true) {
-      Map<String, dynamic>? meta = jsonFormat(result['meta']);
-      return meta ?? Map();
+      return jsonFormat(result['meta']) ?? Map();
     }
     return Map();
   }
@@ -457,8 +491,10 @@ class TopicCommon with Tag {
 
   Future<bool> delete(int? topicId, {bool notify = false}) async {
     if (topicId == null || topicId == 0) return false;
+    TopicSchema? topic = await query(topicId);
+    if (topic == null) return false;
     bool success = await _topicStorage.delete(topicId);
-    if (success && notify) _deleteSink.add(topicId);
+    if (success && notify) _deleteSink.add(topic.topic);
     return success;
   }
 
@@ -482,6 +518,7 @@ class TopicCommon with Tag {
       joined,
       subscribeAt: subscribeAt ?? DateTime.now().millisecondsSinceEpoch,
       expireBlockHeight: expireBlockHeight,
+      createAt: DateTime.now().millisecondsSinceEpoch,
     );
     if (success && notify) queryAndNotify(topicId);
     return success;
